requirements.txt
Flask==2.3.6
stripe==11.38.0
python-dotenv==1.0.0
sendgrid==6.12.0
requests==2.31.0
